+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 20.12.0
 - First major release
+- Bugfixes
+- New webview plugin for NC login
 
 ## 0.9.12
 - Dark Theme & Amoled Mode